requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
langchain==0.1.0
langchain-core==0.1.10
langchain-text-splitters==0.0.1
langchain-chroma==0.1.0
langchain-google-genai==0.0.5
google-generativeai==0.3.2
beautifulsoup4==4.12.2
pypdf==3.17.4
ebooklib==0.18.2
chromadb==0.4.18